sp_event_handle: New public typedef for the OS handle type
Introduce a typedef for the host system's event handle type. This
was already implicitly part the ABI, but library users had to apply
exactly matching casts themselves.

On Windows, sp_event_handle is a typedef for void*, in order to
avoid pulling in Windows system headers in libserialport.h. This
matches the corresponding typedef from winnt.h in STRICT mode.

Use the new type in the declaration of sp_event_set::handles, so
that user code can avoid casting the array. This change does not
affect ABI, and at least for C code API compatibility is also
maintained.